requirements.txt
streamlit==1.31.0
openai==1.12.0
anthropic==0.18.1
python-dotenv==1.0.1
qrcode[pil]==7.4.2
Pillow==10.2.0
python-dateutil==2.8.2